// frontend/src/client/types.gen.ts
export type PortfolioCreate = {
    name: string;
    description?: (string | null);
    is_default?: boolean;
    is_active?: boolean;
};

export type PortfolioPositionPublic = {
    quantity: number;
    average_price: string;
    total_investment: string;
    current_value: string;
    unrealized_pnl: string;
    unrealized_pnl_percent: string;
    id: string;
    portfolio_id: string;
    stock_id: string;
    last_updated: string;
};

export type PortfolioPublic = {
    name: string;
    description?: (string | null);
    is_default?: boolean;
    is_active?: boolean;
    id: string;
    user_id: string;
    created_at: string;
    updated_at: string;
};

export type PortfolioSummary = {
    portfolio_id: string;
    portfolio_name: string;
    total_investment: string;
    current_value: string;
    unrealized_pnl: string;
    unrealized_pnl_percent: string;
    total_positions: number;
    positions: Array<PortfolioPositionPublic>;
};

export type PortfolioUpdate = {
    name?: (string | null);
    description?: (string | null);
    is_default?: (boolean | null);
    is_active?: (boolean | null);
};